The instruction `lpb` (loop begin). Made the loop range parameter more future proof, by replacing u8 with u64 and checks for extreme values.

+++ rust_project/loda-rust-core/src/execute/node_loop_constant.rs
@@ -2,13 +2,12 @@ use super::{EvalError, Node, NodeLoopLimit, Program, ProgramCache, ProgramSerial
 
 pub struct NodeLoopConstant {
     register_start: RegisterIndex,
-    range_length: u8,
+    range_length: u64,
     program: Program,
 }
 
 impl NodeLoopConstant {
-    pub fn new(register_start: RegisterIndex, range_length: u8, program: Program) -> Self {
-        //panic!("TODO: replace u8 addresses with u64");
+    pub fn new(register_start: RegisterIndex, range_length: u64, program: Program) -> Self {
         Self {
             register_start: register_start,
             range_length: range_length,
